Update Leakage Definition in lcm_skeleton.als
We expect to have e3 to be an intevening access (interfering access) that disrupts the extra-architectural flow of information between e1 and e2. If so, we require (e3--[ecomx]-->e2) rather than saying that (e3--[ecomx]-->e1). Additionally, why do we not have explicit statements that constrain e1, e2 and e3 are events that access the same xstate-location?